beging gerrithub.io integration
Redesign the data structure to be able to pin on a given git version or
a componant.

TODO(Gonéri): I think we should move the JSON part that describe the
product repository in a sub components section:

data/components/dci-control-server/git
data/components/dci-control-server/rev

I will work on that tomorrow.

Change-Id: I9959dabb63ad6576464992837ab0e97400a5b38d