beging gerrithub.io integration
Redesign the data structure to be able to pin on a given git version or
a componant.

I think we should later move the JSON part that describe the
product repository in a sub components section. For example:

- data/repositories/dci-control-server/git
- data/repositories/dci-control-server/rev

And the test configuration should be in another section. Like:

- data/tests/tox
- data/tests/khaleesi

This way, it will be easier to see the difference between the
environment that has to be prepared before the test and the test
configuration.

Change-Id: I9959dabb63ad6576464992837ab0e97400a5b38d