products: ability to associate different repo
So far a product was a repository. This was working fine for basic
product but for complex product this is too much limited.
We now use a components dictionnary to list the repositories associated
to a product.

So far, the components structure is embedded in the data structure. But
we will probably create a new table later.

Change-Id: I52c451dbf84ebf1513f12a1e61fe2ea7300de699