Add posargs possibility in tox
According to tox documentation it is possible to define a default
command line if no posargs are provided to tox.

This patch allow the developer to customize the tox command call in
order to pass arguments to the underlying command.

Example: tox -epy27 -- -q server
Which will run the tests silently only for the server directory.

Change-Id: I0c358f4f9cea4abf8cc5a4c486fa8425ce30ce20