UI: show version details
Partial support for /versions and /versiondetails. We still need to
properly unfold the result associated to a version.

I the begining I was expected to be able to rely on Python-Eve embedded
items to collect automatically the sub-items:

version:
  - testversions
    - test
    - jobs
      - jobstates

But we need to improve the Python-Eve code to be able to do so.

Change-Id: I79c22a4384a96b0e6e609f84e086695d1e614069