Avoid db collision on process failure
If the test process fails, the system can keep
an outdated database which will collide with
the new one created on each test session.

This patch ensure that the db is deleted at test
startup if there is a remaining one before populating
the new one.

Force app creation on function scope instead of session.
This will be useful for further testing because it avoids
tests to modificate the app object and cause some side
effects.

Change-Id: Ia846389f9af104871d7c272b9d9876ae08754865