find_or_create_or_refresh should update with PATCH
PUT totally destroy the existing resource and break relationship with
the other resources. This is not what we want. Let's use PATCH instead
here.

This patch ensure we make use of datetime _without_ timezone, this
because datetime datetime from Py27 is not able to compare them properly
and this create error in the server:

    can't compare offset-naive and offset-aware datetimes

offset-naive is the newly generated `updated_at` value whereas
offset-aware is the value from the DB.

Change-Id: I62bc32b18170816ac36cf7c22c62c0f6feb0e365