Fix moment objects instanciation
Moment object where instanciated using the javascript object Date
which is not the way recommended by the library.

According to this, the string pattern is used instead of Date object,
the moment instanciation is also updated to strict mode in order to
speed up instanciation and avoid multiple date format.

The invalidDate string has also been updated in fr, and en locales
in order to display "N/A" instead of "Invalid Date"

Change-Id: I6ac71268481d5e8062e5b886c69898edd014912b