Update patch set 11
Patch Set 11:

LIKE is not planned to be implemented until we need it. I didn't see an agent using it.

We do not need to support the NOT operator since we planned to support the others:  !=, <, > etc...   not(a==b) is a != b

We already support the 'embed' for the join.

Patch-set: 11
Reviewer: Gerrit User 29 <29@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>
Label: Code-Review=0