Remove list from the validation error
The validation errors field was a list at first

ie: errors: {id: ['id field required']}

This was due to the fact that multiple errors can happen on a same field.
I based this feature on the marshmallow library but voluptuous was chosen,
and voluptuous does not support this feature, only one error is possible
on a field. According to that, the list of errors is no longer needed and
has been removed in this patch.

Change-Id: Ic253ceb7b548745bea5a4b916f6afaa4b0bc8309