Update patch set 2
Patch Set 2: Code-Review+1

As an end-user it works as expected. Run an agent, upload file, and the content is indexed and queriable in the ES index. Thanks Cedric !

As for the approach, I am not sure yet - MaxV might have better insight - but I would push to the direction where we can merge this as soon is it is compliant to our guidelines, and use it, this is when we will find out what are our use-cases and what needs to be adapted.

Patch-set: 2
Label: Code-Review=+1