Add a test debug capability in gulp
Debugging test on protractor + phantomjs was complicated due to the lack of UI.
This patch add a test:e2e:debug target in the gulpfile which will allow
the user to launch a chrome to debug tests.
The compose and Dockerfile have been modified accordingly in order
to install google-chrome, google-chrome-driver and share the host display.

Change-Id: I5c19dda19577ef83b1e4a6bd7f1cc84bbb7b0b0f