Add 'status' field to job resource
Currently it's very slow to get the status of all the jobs,
it's a N+1 select issue because for each job we must first
get the associated jobstates in order to get the last one.

This commits solve the issue by adding a 'status' field
inside the job table. After each creation of a new jobstate
then the associated job status is updated according to the
new status.

Change-Id: I636816277b4c015646b0dd95413d29fd0a6df722