Fix issue #8 and #5
This patch fix the feedback when clicking the recheck button
in the UI: https://github.com/redhat-cip/dci-control-server/issues/8.

It also fix the scrollbar on file display with a word-wrap:
https://github.com/redhat-cip/dci-control-server/issues/5

Now the button change color when is hovered and have
a tooltip to explain what the button will do, i.e: recheck the job.
It also display a minus-sign and is disabled when a job is under recheck

This patch also include some refactoring for a better UI.

    - The issue which was logging a bunch of error in the console
    has been fixed. It was due to the redirection of the index state.
    Now the index state reject the resolution with a 301 which is handle
    by the app and redirect to the /jobs page

    - Bootstrap sass is now used and integrated in the toolchain, this
    allows the developper to access some variables in the dashboard.scss
    file.

    - The custom css has been removed, we now rely on bootstrap for the
    UI design, this is due to the lack of design skills in the team.

    - An issue in the recheck controller has been fixed, this issue caused
    recheck job to start with a non new status, due to the copy of the old
    job object.

    - Now details of a jobstate are accessible through a "see details" button
    The first jobstate of a job which have files to display is openened by
    default.

Change-Id: I5780a4b573bc69bc10213ca0e04e54c27df1f2e8