Remove docker compose specific entries
The docker compose settings does not belong here because
they are specific to the development environment.
See here: https://github.com/redhat-cip/dci-dev-env
for the development environment.

This patch remove those settings and provide the most simple
settings. Some hooks are available to keep this repo running
in the dev env.

Change-Id: I41684dcb67af7e4c52c552d88c7152df69311afc