db: Add a job_id column in the files database/
Users wants to be able to upload context file (be it in .ini, .json or
any other text format) and tight them with a job.

This commit sets up the building block of this feature by adding a new
column in the files table that will link files and jobs together.

jobstate_id becomes nullable as attaching a context file does not need a
jobstate_id

Change-Id: I67d622095964306d6dedfd90cb7b31663cfa3a5e