[Bugfix]: fix user creation
Currently we use a composite unique constraints on the couple
(user.name, user.team_id) in order to make possible to have
two users with the same name on different teams.

This is messy and leads to the issue
https://github.com/redhat-cip/dci-control-server/issues/25

Because in the authentication scheme we rely on the user name
to get its infos, this can leads to retrieve the wrong user
from the database.

This commits fix this issue by forcing a unique name for a
user among the whole teams. Thus it ensure that the retrieved
user in the authentication scheme is the right one.

Change-Id: I91fc351ced16d49101d70cc4a87bed32f4576d38