set MAX_CONTENT_LENGTH to 20MB
At this point, we do not limit the size of the POST. At this point,
Control Server is not designed to handle large amount of data.

Full document data are stored in memory then injected in a new row
in the DB. Later, we will collect all these row and present them
in the dashboard. We can hardly accept file larger than 20MB because
of this.

Change-Id: I800af930ff433d269c5b333cbb0d2a2e336d9f0e