InfluxDB: Create a new entry per job scheduled
In order to collect metrics on a job basis, everytime a job is scheduled
we create a new database within influxdb. A new user is created if
necessary (ie. one user per tenant). The credentials and the IP/PORT
setup informations are returned to the agent so the agent can send
metrics directly to the influxdb instance.

Change-Id: I51d1822004ad6d06a2b398d6642e5a4e81f2bc55