jobdefinition: add active and comment columns
active
------

We use it to ensure a jobdefinition won't be scheduled. At this point
the default value is: `True` for backward compatibility. We should
quickly change this to `False` and let the API user enable it when the
jobdefinition has all its components ready.

comment
-------

Allow to add some comments regarding a given version, for example it may
be interesting to make explicit later after a release, that a jobdefinition
is actually the final release.

Change-Id: I14a862744681744f10b7ed24311a8ae12c449b91