JUnit: Handle Junit files
Currently when junit results of a job are submitted they are treated as
any other file. They're displayed in raw format. Making them not easy to
read.

The purpose of this review is to serialize JUnit files into python
object that the API sends back to the UI. The UI is then in charge to
display them nicely.

Change-Id: I35d953258611389d883547430d6853bec199ae5f