schemas: Fix jobdefinition default when submitting a PUT
The jobdefinition resource as default value for active and comment. This
is fine when we receive a POST request but it becomes problematic when
we receive a PUT request as it overides previous values.

Scenario:

  1. Create a jobdefinition resource. It will have a default value for
active of True
  2. Set active for this jobdefinition to false.
  3. Set the comment for the jobdefinition to 'A comment'
  4. Make a jobdefinition show of the jobdefinition resource, the active
value is back to True.

This is because no 'active' field has been detected in the payload then
the schemas provided its default value.

Change-Id: If8e06033bc31ef72c52821ad73ceff2c1d781e12