files: streaming files
This commits allows to:

    - upload a file with a streaming fashion
    - download a file with a streaming fashion

in order to avoid memory issues.

The metadatas are now provided through http header
while the data is the content of the file.

The content of the file are now stored in the local
FS, in the directory specified by FILES_UPLOAD_FOLDER.

When uploading a file, the control server will store it in
the path:  FILES_UPLOAD_FOLDER/<team-id>/<file-id>.

NOTE: The tests in test_search.py has been deactivated because
we don't push anymore the content of the files from conftest.py
this is done asychronously. We need a new way to test it.

Change-Id: I8422875f717bca7558d42cc4c566bf113a9a54de