jobstates: Retrieve file content allong all jobstates
Currently we retrieve a list of jobstates and then retrieve the files
attach to this jobstate, each of this retrieval requires a new request.

So assuming we have 50 jobstates for a given job. To be able to retrieve
all the information we need currently requires 51 HTTP queries.

This commit aims to ease that allowing a user to embed the file
parameter and hence retrieve all the content in a single query.

Change-Id: I7fdd8e8efaaa351638ff890260ac06c9a42011b1