fix the file migration
The file migration uses the API. The API code expect the size column to
be available. But this column is added by the next migration script.

This idea is to create the size column before the migration. This is a
problem we are likely to encounter in the future. We need to adjust the
script to avoid the use of the API.

The commit also resolve a little UTF-8 issue.

Change-Id: Ibb7deb3c9efdf7e47be4a8e67c8d8beae9559aee