WIP jobs: Remove configuration from the object returned
TODO: UNIT-TEST is broken and we need a test coverage.

Currently when querying GET /jobs, all the output from
tripleo-stack-dump is retrieved for all jobs in the first page,

This result in an unecessary heavy workload for something that is not
displayed. So when retrieving all /jobs informations we remove the
configuration part of it.

Change-Id: Ia517a4ddc2b0c687dbefbf4f6ec39734d26d4539