drop jobdefinitions.type
So far, the scheduler was using the topic_id parameter: topic_id to
specify the type of job wanted.
The new scheduler2[0] comes with another parameter called type to give
the ability to use some extra parameters for the installer.

Is new parameter is not really necessary:
At this point, the users need to specify the topic_id when they run the
agent. With this type parameter, we will create a new mandatory
parameter. At the end, the configuration will be more complex.
So far, we have:
- the topic to specify the inputs stream
- the configuration field to describe the installation result
- the remote CI to describe the infrastructure and its configuration

This is enough for our current partners and there is no user story asking
for a better granularity.

IMO we should remove it from the scheduler2 and only relies on the
topic_id instead. As a bonus point, this will make the scheduler2
compatible with scheduler.

[0] I3e7abe2d626209ca63101ab50009ee4b13cb2d9b

Change-Id: I19ca4c5de55a25378ce6913eacdec4bfe32fb193