junit: Fix /jobs/<j_id>/results endpoint with empty result
When a junit test suite has been updated with an empty run, the skips
parameter is missing. Thise caused the /jobs/<j_id>/results endpoint to
be broken.

In such case, we test now if it is missing, and if that is the case we
set it to 0, so the rest of the method can work smoothly.

Change-Id: Ibb6df4f267b033a63a7ff7e6c1ee1b3c4e918278