Move tests to belongs to team.
In this way team member can create tests and attach them to element.

User can:
 - Create a test
 - Attach a test to a remoteci
 - Get tests from remoteci
 - Get Tests from topic

Only the admin can :
 - Attach a test to a topic

All tests belongs to team so if a user is not in the team's tests he
can't interact with it. The only exception is for topic. If a user is in
a topic he can get all tests, even if the test does not belongs to his
team.

So the user can create specific test for his remoteci and can always get
the generic test for the topic.

Change-Id: I6570c4a72391499b8c3051fa1d7d5c165c016faf