access-control: no restriction for admin user
The admin team is managed by us and so should now have export control
restriction.

In order to be able to identify the admin user, we now retrieve the user
team_name when we collect user details. Because of that, we don't need
anymore to initialize the TEAM_ADMIN_ID key during the application
start up.

Change-Id: I8409690ab0184b63df348c252c341120ccf88ee5