jobs: ability to embed component+fix many embed
This commit fix the sub-embed like the follwing ones for jobs:
- jobdefinition.tests
- remoteci.tests

And add this new embed to job:
- components

To component:
- files

To team:
- topics

Overall, the patch enforce the use of dedup_rows() (was parse_rows()) as much
as possible. Without this, the use of a "many=True" embeds will only return
a dict of the last entry instead of the full list.

One record information may be splitted on more than just one SQL row. These rows
will be ultimately merged in one line by dedup_rows().

E.g:

  | id   | embeded_field
  | id_1 |         data1
  | id_1 |         data2

If we use a limit operator here, we may end up with a truncated record.
This is the reason why we do a first iteration to only identify the IDs.
The second iterations is used to actually collect the full record
data.

Change-Id: Ia0f84bc97ef459e2ff551fb380f04e2363c2426d