jobs: ability to embed component+fix many embed
This commit fix the sub-embed like the follwing ones for jobs:
- jobdefinition.tests
- remoteci.tests

And add this new embed to job:
- components

Overall, the patch enforce the use of dedup_rows() (was parse_rows()) as much
as possible. Without this, the use of a "many=True" embeds will only return
a dict of the last entry instead of the full list.

Change-Id: Ia0f84bc97ef459e2ff551fb380f04e2363c2426d