Do not hard-delete resources anymore
Currently when one DELETE a resource, the resource is deleted from the
server and all the ON_CASCADE action are triggered.

This, while working, can result in severe data-loss if the user inputed
the wrong ID, as there is no safe-net. This commit aims to soft-delete
the resources in the server, by changing their status from Active to
Archived. Archived resources won't be retrieved when listing all
resources.

Later on, the dcictl purge command can be run to actually delete
Archived items.

Change-Id: I6fa7110a3c1e55598c829597c14821167c9ab953