embed: ability to embed two time the same table
This commit gives the ability to embed two time the same column, in this
case we have to use an alias() each times. Previously this was not
something we were able to do because all the join complexity was hiden
in v1_utils.embed().

To get the status of a remoteci, we can know do thinkgs like:
/remotecis?embed=last_job,last_job.components,current_job,current_job.components

So far v1_utils.embed() was enforcing the use of a one-level outerjoin.
We know build the class to join manually and pass it to embed() through
the join parameter.

This commit also:
- add the ability to use embed in most of the GET view
- drop flatten_columns(), we don't need it anymore when we use use_labels=True.
- simplify get_jobs_status_from_components()

Change-Id: I6eafe48ed14ff4835495e2dfcc55ff3c8fcb7433