WIP: embed: ability to embed two time the same table
This commit gives the ability to embed two time the same column, in this
case we have to use an alias() each times. Previously this was not
something we were able to do because all the join complexity was hiden
in v1_utils.embed.

To get the status of a remoteci, we can know do thinkgs like:
/remotecis?embed=last_job,last_job.components,current_job,current_job.components

This commit also:
- rename the components embed from files to component_files to be
  consistent with the name of the object returned by the
  /components/<c_id>/files route. We may want to do the opposite?
- drop flatten_columns(), we don't need it anymore when we use use_labels=True.
- MORE TO COME

Change-Id: I6eafe48ed14ff4835495e2dfcc55ff3c8fcb7433