zmq: use a global variable to preserve the context
The DciControlServer class is recreated before every unit-test run.
Sometime, we face the situation where the ZMQ context deletion was
blocked in Context.term(). This was freezing the whole test process.

The ZMQ Context is associated to the ZMQ Socket and since we don't
have the need to recreate it all the time, we now trace it in a global
variable. This way, we never delete the ZMQ Context during the test run
and we speed up the DciControlServer instance creation.

Change-Id: Ifeaa5139b1eec6cf2f7139c942add03c5e5c2bd1