Update patch set 3
Patch Set 3:

Just to clarify: we can already wipe large part of the DB just with some DELETE, because of the cascade on delete. This is not a regression introduced by this patch. However, the verge purge sounds like the operation is safe, which is absolutely not the case until we drop the cascade on delete :D.

Patch-set: 3