Update patch set 2
Patch Set 2: Code-Review-1

I agree with Dimitri here. I don't get the need for this patch. I don't see when we might purge more than expected.

The 'state' patch was a protection against this. But the cascade on delete rules are there only to not have left overs.

Let's say you delete a TEAM, its normal to remove all the link between this team and the various topics it's been attached to, remove the tests that was attached to this team the same goes for the user. This is not data loss, this is stating two times once for soft-delete and once for hard-delete that you want to remove an item.

Plus in case of emergency we still have the backup if an admin stated two times s/he wanted to remove an item.

Patch-set: 2
Reviewer: Gerrit User 115 <115@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>
Label: Code-Review=-1