Update patch set 2
Patch Set 2:

To me 'active' means 'not deleted'. This is my problem.

As a user, I expect the purge to only purge what is deleted. This is not the case in your example. E.g: I delete a topic, 2 weeks after I start to debug one of the jobs associated to the topic, someone purge the topic, I lose my job even if it was 'active'.

- we can either 'delete' all the sub-resources when we delete something (by adding some extra Python code)
- or prevent the purge to do remove sub-dependencies.

Patch-set: 2