Update patch set 2
Patch Set 2:

> I start to debug one of the jobs associated to the topic, someone purge the topic, I lose my job even if it was 'active'.

Yes and that the expected output. If as an admin I purge a topic, there is a good reason for it, and all attached items should be removed. I will never purge accidentally, however I might archive it accidentally.

The only thing I see, is that today with the new state feature, we don't follow what an delete would do.


> I would be perfectly fine with the cascade on delete if, when we delete a team, we also delete all the sub-resources.

So my understanding of this sentence becomes: if we archive a team, all the resource that would be deleted with the CASCADE rules should be also updated to state 'archived'.

Is that correct? In which case I agree that one more patch should be added on the control-server side to mimic the CASCADE rule.

Patch-set: 2