topic: don't return only one job in status
During the left join between remotecis and jobs we are using a
limit 1 on jobs.
This results to have only one job informations available even if we
have multiple remotecis with a valid job on that topic.

remotecis | job id
----------+--------
 remote A |  null   // never run
 remote B |  null   // run at 17:30
 remote C |  XXXX   // run at 17:35
 remote D |  null   // run at 17:15

It's the most recent job that is display in this case.
This patch allows to fix the issue.

remotecis | job id
----------+--------
 remote A |  null   // never run
 remote B |  XXXX   // run at 17:30
 remote C |  XXXX   // run at 17:35
 remote D |  XXXX   // run at 17:15

Change-Id: Ib6112d46fdb655847c63735b413bea376935cb00