Mutualize in a single method the base values dict
Most resources we create relies on the same basic informations (id,
created_at, updated_at, etag). We, currently, redo this operation in each
resource, making it error prone and not using the DRY pattern.

This review aims to mutualize the part where this basic dict is created,
and then each create_RESOURCE method will call it. Hence reducing the
amount of places where the code needs to be changed when necessary.

Change-Id: I7451ce37a9897e7e699628143393aa6f358b2480