jobdefinitions: Fix topic id filter
Since 27cfa2d78aaec7376a82c0bc1028db53ff40fc32 the endpoint
/topics/<uuid>/jobdefinitions is not working anymore.
As an admin it was returning all jobdefinition instead of those for
a specific topic.
As an user it was returning as python trace:
  TypeError: 'UUID' object is not iterable

On this particular endpoint the topic_ids is not a list so we can't
use the IN sql operator.

Change-Id: If4474f5b5be567b8886b2734f963e9e8b92046e7