Role: Initial commit for the role resource
This commit introduces the roles resource.

This commit does not implement the usage of the mechanism itself but
really implement the ground to be able to use it. roles will be
resources that will hold permissions and that will be attached to users.

Admin of the plateform can create roles, so can the admin of each team.
Team can have their own roles.

Change-Id: I45500a2b4f2a3997dcc055ae9d24f869afa01724