files_events: Delete FK on files.id
The files_events table references the files table via file_id column with a
foreign key. But the goal of the files_events table is to track the events so
we don't need to have a reference on files because we want to keep the event
list even if the files is deleted.

Change-Id: Ia51858ff7aa8c24e250e4ff343de8001e608a76f
Closes #30