Update patch set 1
Patch Set 1:

I don't think we need to use a DELETE ON CASCADE here. The purpose of the files_events is to trace the changes to be able to sync ElasticSearch. If we remove a file, we should add an event saying the file has been deleted.

So maybe we don't need a FK here?

Patch-set: 1