job create: team_id not mandatory
Since the remoteci user is already associated to a team, there is no reason
to enforce the team_id parameter during the creation of a job.

The parameter is still interesting if the admin want to create a job for
a specific team.

As a side effect, the patch fix a bug that was allowing a user to create
job for another team.

Change-Id: I6bd810ea8572dfec1bea41c1369ef0fd222b7cc1