[start_es]: fix start_es.sh race condition
Currently the script ./scripts/start_es.sh use
netstat and grep the port to determine if
elasticsearch is running. But this is not sufficient
since we run elasticsearch as a daemon with the -d
option and having it listening on the port does not
mean it's fully running.

This lead to some race conditions in the tests. This
patch replace the netstat test by a call api with curl.

Change-Id: I92af74ff1ddbd84e720c6adc6b9bb574bd69842c