Update patch set 4
Patch Set 4: Code-Review-1

(2 comments)

I understand that the goal is to have a remoteci configuration but I wonder why we should just go with /configurations ?
What I don't understand is why this should be under /remotecis when a config is also linked to a topic too. But the implementation is different. One (topic ref) is a FK in rconfig table and the other (remoteci ref) is an entry in the join table. Why not using the join table with a tuple ? ex: (config_id, remoteci_id, topic_id)
Finally the purge mechanism is not implemented.

Patch-set: 4
Label: Code-Review=-1