Update patch set 1
Patch Set 1:

Maybe we can bring this discussion during a daily?
But I also asked Yanis why we obliged to add fullname.

The reason was that it is interesting to be able to send an email to the person by calling it by his first name rather than his login.

I'm no sure where the domain control should belong. If we got an API simpler to use, maybe we should enforce only username or email. To reflect use case described, maybe the interface can force full name. But only the user interface, not the API.

Patch-set: 1