Update patch set 1
Patch Set 1:

>  the user may have a good reason to want to keep the fields blanks

Can you tell me a valid use-case where a user wouldn't want to specify his professional email address since this tool his aimed to our professional partners/customers and not to individuals.

So, fullname can default to name, so his nullable/non-nullable can be abstracted to end-user. But I don't see why email should not be non-nullable.

Yes, it is not mandatory for DCI and user to works per-se, but IMO, it is mandatory to have a real plateform. Let me mention some example:


* As an administrator I create a new user, I've been provided with his name and his email address before creating it. After creating his user in DCI, the partner receives en "activation' email when he can set his own password and confirm that the email is valid


* As a DCI user I'd like to reset my password since I've forgotten what it was. On distributed-ci.io I click Forgotten Password link and ask for the regeneration process that will be send by email


* As a DCI user I'd like to be notified on job failures


More use-cases in a the future, could lead to user receiving a montlhly health summary of the various topics jobs status, people being able to attach one to a job so he can take a look at it, etc...


So yes, it is not mandatory, but most SaaS platforms require it (ex. Github, TWitter, Trello, ...) even though there is no strict need for it for their platform to work. The email addresse allows a platform owner to be able to add new features that could rely on it.

Finally, once we will have the SSO plugged in, email will be present. This is just a step closer to try to have the same user schema as SSO would provide us with.

Patch-set: 1
Label: Code-Review=0