Update patch set 6
Patch Set 6:

In fact, i did the PUT on patchset 3. But after talking with Gonéri, its not a good idea to allow PUT on rconfigurations in case they are already used by some jobs, this could lead to inconsistencies.

So instead to update it, one should create a new one.

Patch-set: 6
Label: Code-Review=0