Don't stream files when using signature authentication
When using signature authentication, the WSGI.INPUT stream
`flask.request.stream` has already been consumed when calling
`flask.request.data` when computing the sha256 hash for the content of
the request.

As a general case, if the stream as already been consumed, the data is
already in memory in `flask.request.data` and can be passed to swift
instead.

Change-Id: Ic0a03ec19029d06bba0e8c8ba7becc09549dd129