Update patch set 5
Patch Set 5:

We all had the case where we want to quickly test something with a new account. So we enter
whatever we want in a form and boom, we have our new user. Off course, the mail notifications
won't be sent and the user name being empty, the UI will use the login instead. But who are
we to decided for the user if this is ok or not.

Patch-set: 5
Reviewer: Gerrit User 30 <30@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>
Label: Code-Review=0