Junit: Add support for testsuites in our parser
Currently our parse suppose that the root element in a testsuite.
Since this isn't always true, it fails when the root element is a
testsuites (a set of testsuite)

This commit aims to add support for our Junit parse to handle both cases
when root element is testsuite or testsuites.

Change-Id: Id333299c07769224933cd6ce9b398e17869e9a9b