RemoteCI: Introduce the notion of public
This commit allows one to state that one wants the result of a given
RemoteCI to be publicly available, no login_required in order to see the
result.

This commit simply prepare the ground at the database level. Next, a
commit will change the way data are retrieved from the API point of
view.

Change-Id: If5ace37e326d6f79ae075d4378944ccfbab630af