remoteci/data: able to delete existing keys
The user should be able to remote an existing key from the data field.
Until now, it was just possible to extend the key list. This behaviour
was specific to the remoteci, the other resources do not have the
problem.

To clarify, a mechanizm (key_to_remove) was inplace to allow one to
delete keys. However, this was counter-intuitive because the behaviour
was specific to this end-point and to this key.

Closes #41

Change-Id: Ib390cbd2bfe29ea47119a6abbe80e62c592b99b3