API: Externalize get_all_teams()
Within the API code we should be able to retrieve all the teams for a
given user. This needs to be available with both the REST endpoint and
programmatically, hence the core of the function has been externalized.

Also, get_resource_by_id() is filtered by team so resources cannot leak.
However a PRODUCT_OWNER should be able to see 'all' the resources within
its scope and not just the one assigned to its team.

Change-Id: I016540f32e55420f71e3989339787e84d67bbfe4