sso: add sso auth mechanism
The user can now authenticate through
Keycloak SSO server with OpenID connect.

The HTTP header to user is:

    Authorization: Bearer <token>

At the very first login with SSO, the user
doesn't exist in the database. Then it is
automatically created with its SSO information.

At this step the user does not belongs to a team,
this operation is manual and will be done by a
DCI administrator.

Also, to map the dci's users with the SSO's users
we use the SSO username.

Change-Id: I33ca6e4f6e865356c09d9c13ee3dc371170930b3