Identity: Introduce the notion of identity
Currently, when dealing with identity, we only have access to the data
the SQL request returns. Identity related methods are then scared across
differents files (auth.is_admin(), auth.is_in_team(), etc...)

The idea of this method is to keep having access to the SQL returned
data but yet to be able to add helper directly into the class itself.

Change-Id: Ib88a90c3217f04345488e93091c3d81c9de811a4