Update patch set 9
Patch Set 9:

> I don't like the idea a PO can manage components.

On the long run, once the patches will be merged, PO will rely on the Feeder to upload them. As of today it is not possible yet. But a PO is the owner of its product, so I don't see why the owner of a product wouldn't be able to upload its own product if wanted.
 
> What is the purpose of those components.
> We should control the components we offer because we want to compute some stats on it.

You are the PO of OpenStack, you upload puddles. Purpose of the puddles is to test the latest official version of Red Hat OpenStack. You are the PO of Ansible, you upload Ansible puddle of whatever its name is.


> IMO only feeder or super_admin can manage components

Feeder is a resource that doesn't exist yet. Today, if we want the product team to upload their "puddle" for their own product we must share the super_admin account credential. Its insa(n|f)e. This is a first step toward not sharing those. Then once the Feeder will be merged and effective, PO will be able to use the feeders for that purpose rather than their personnal credentials.

Patch-set: 9
Reviewer: Gerrit User 115 <115@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>
Label: Code-Review=0