Update patch set 9
Patch Set 9:

> Feeder is a resource that doesn't exist yet. Today, if we want the product team to upload their "puddle" for their own product we must share the super_admin account credential. Its insa(n|f)e. This is a first step toward not sharing those. Then once the Feeder will be merged and effective, PO will be able to use the feeders for that purpose rather than their personnal credentials.

We are close to have the feeder working. So I would prefer to spend some time finishing the feeder.
I'm pretty sure that when the feeder will be in place the PO will still be able to CRUD components. So we are adding complexity now on components controllers. IMO this is not a step forward.

Patch-set: 9