Update patch set 1
Patch Set 1: Code-Review-1

Whats the different between internal and external remotecis?

If this is for remotecis to not show up on the UI, shouldn't be this flag put on the team level ? 

I have my own team, yaya has is own team, dsavinea has is own team, etc...

Then all the rci from a team would be either internal or external. All this comment depends on the answer to first question.

Patch-set: 1
Label: Code-Review=-1