Update patch set 1
Patch Set 1:

> Can a team have 2 remotecis, an internal and an external?

I think the term external is poorly chose here. Because, I'd like to say a team cannot have internal and external CIs. Yet a team can have RemoteCI they want to count in stats and RemoteCI they don't.

Yet, if we go down this road, we have no way to enforce partner's RemoteCI to all show up in stats if for some reasons they decide they don't want them to show up. Then, we'll be left with no data making DCI not as usefull as we'd like.

So to me we have internal and external 'Teams', internal might be used for tests, debug, etc... and you don't want them to appear in stats which is fine.

For partner team I think all their remoteci should count in stats, its the only way to ensure we gather as many data as possible.

Patch-set: 1