Update patch set 1
Patch Set 1:

We can have a partner with an experimental remoteci. Having an explict way to ignore the results coming from this remoteci will be useful. For the same reason we may have an internal lab that return valuable information that we want to show up in our stats.

internal/external is too specific to our current context. What really mater is: Should this remoteci results be aggregated in our stats.

Patch-set: 1