Update patch set 13
Patch Set 13: Code-Review-1

(1 comment)

I don't like this one:
We are mixing persistence with the domain. Identity should be an object with properties and methods. We should use polymorphism or composition to create an Identity. Identity should be tested apart, with some unit tests.


Let me try to explain with more details. If Identity.teams depends on user's role, then the domain should be in the constructor of Identity.

Identity(user, role, teams)

Base auth mechanism should only get all teams in the db, then create an Identity object, filtering the teams based on user's role in the constructor.

We should not introduce domain in the SQL alchemy constraint object.
This is an anti-pattern, and it will be really difficul to maintain.
Thus we cannot test this domain without a working database.

We can maybe do a pair programming live stream on bluejeans if you want, because I think we are going in the wrong direction.

Patch-set: 13