Jobs: Remove the explicit need to specify RemoteCI
Currently in order to schedule a Job we rely on an explicit field in the
payload called remoteci.

Since remoteci should be the only role allowed to schedule a job, we already
have this information made available. Hence we re-use this information
if available but keep the previous behaviour in place for backwards
compatibility until proper role verification is put in place.

Change-Id: I9a45b648feae05a229544e85c25e10a343505d86
Depends-On: I61e6c5432940f655ea1235095af597604b4a9a85